@@ -13,78 +13,76 @@
     "lint:openapi": "openapi lint ./openapi3.yaml",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
-    "release": "standard-version",
-    "test": "npm run test:unit && npm run test:integration",
+    "test": "tsc --noEmit -p tsconfig.test.json && npm run test:unit && npm run test:integration",
     "prebuild": "npm run clean",
-    "build": "tsc --project tsconfig.build.json && npm run assets:copy",
-    "start": "npm run build && cd dist && node ./index.js",
+    "build": "tsc --project tsconfig.build.json && tsc-alias -p tsconfig.build.json && npm run assets:copy",
+    "start": "npm run build && cd dist && node --import ./instrumentation.mjs ./index.js",
+    "start:dev": "npm run build && cd dist && cross-env CONFIG_OFFLINE_MODE=true node --enable-source-maps --import ./instrumentation.mjs ./index.js",
     "assets:copy": "copyfiles -f ./config/* ./dist/config && copyfiles -f ./openapi3.yaml ./dist/ && copyfiles ./package.json dist",
-    "clean": "rimraf dist"
+    "clean": "rimraf dist",
+    "generate:openapi-types": "openapi-helpers ./openapi3.yaml ./src/openapi.d.ts --format --add-typed-request-handler",
+    "prepare": "node .husky/install.mjs"
   },
   "directories": {
     "test": "tests"
   },
   "author": "MapColonies",
   "license": "ISC",
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
-  },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
-  },
   "dependencies": {
-    "@godaddy/terminus": "^4.10.2",
+    "@godaddy/terminus": "^4.12.1",
+    "@map-colonies/cleanup-registry": "1.1.0",
+    "@map-colonies/config": "^1.3.2",
     "@map-colonies/error-express-handler": "^2.1.0",
-    "@map-colonies/express-access-log-middleware": "^1.0.0",
-    "@map-colonies/js-logger": "^0.0.5",
-    "@map-colonies/openapi-express-viewer": "^2.0.1",
-    "@map-colonies/telemetry": "^4.1.0",
-    "@opentelemetry/api": "^1.1.0",
-    "@opentelemetry/api-metrics": "^0.29.0",
+    "@map-colonies/express-access-log-middleware": "^2.0.1",
+    "@map-colonies/js-logger": "^1.0.1",
+    "@map-colonies/openapi-express-viewer": "^3.0.0",
+    "@map-colonies/schemas": "^1.3.0",
+    "@map-colonies/telemetry": "^7.0.1",
+    "@opentelemetry/api": "^1.7.0",
+    "body-parser": "^1.20.3",
+    "body-parser-xml": "^2.0.5",
     "compression": "^1.7.4",
-    "config": "^3.3.7",
-    "express": "^4.18.1",
-    "express-openapi-validator": "^4.13.8",
-    "fast-xml-parser": "^4.0.8",
-    "http-status-codes": "^2.2.0",
-    "reflect-metadata": "^0.1.13",
-    "tsyringe": "^4.7.0"
+    "express": "^4.21.2",
+    "express-openapi-validator": "^5.4.2",
+    "fast-xml-parser": "^4.5.1",
+    "http-status-codes": "^2.3.0",
+    "prom-client": "^15.1.2",
+    "reflect-metadata": "^0.2.2",
+    "tsyringe": "^4.8.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.0.1",
-    "@commitlint/config-conventional": "^17.0.0",
-    "@faker-js/faker": "^7.1.0",
-    "@map-colonies/eslint-config": "^3.1.1",
+    "@commitlint/cli": "^19.7.1",
+    "@commitlint/config-conventional": "^19.7.1",
+    "@faker-js/faker": "^9.4.0",
+    "@map-colonies/eslint-config": "^4.0.0",
     "@map-colonies/node-osm-elements": "^2.1.2",
     "@map-colonies/prettier-config": "0.0.1",
-    "@map-colonies/standard-version-update-helm-version": "^2.0.1",
-    "@map-colonies/standard-version-update-openapi3-version": "0.0.2",
     "@redocly/openapi-cli": "^1.0.0-beta.94",
-    "@types/compression": "^1.7.2",
-    "@types/config": "0.0.41",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^27.5.1",
-    "@types/multer": "^1.4.7",
-    "@types/node": "^17.0.43",
-    "@types/supertest": "^2.0.12",
-    "@types/yamljs": "^0.2.31",
-    "commitlint": "^17.0.1",
+    "@types/body-parser": "^1.19.5",
+    "@types/body-parser-xml": "^2.0.5",
+    "@swc/core": "^1.7.26",
+    "@swc/jest": "^0.2.36",
+    "@types/swagger-ui-express": "^4.1.7",
+    "@types/compression": "^1.7.5",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.14",
+    "@types/multer": "^1.4.12",
+    "@types/node": "^22.13.0",
+    "@types/supertest": "^6.0.2",
+    "@types/yamljs": "^0.2.34",
     "copyfiles": "^2.4.1",
-    "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.16.0",
-    "husky": "^8.0.1",
-    "jest": "^28.1.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.43.0",
+    "husky": "^9.1.7",
+    "jest": "^29.7.0",
     "jest-create-mock-instance": "^2.0.0",
-    "jest-html-reporters": "^3.0.8",
-    "prettier": "^2.6.2",
-    "rimraf": "^3.0.2",
+    "jest-html-reporters": "^3.1.7",
+    "prettier": "^3.4.2",
+    "rimraf": "^6.0.1",
     "standard-version": "^9.5.0",
-    "supertest": "^6.2.3",
-    "ts-jest": "^28.0.3",
-    "typescript": "^4.3.4"
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.2.5",
+    "tsc-alias": "^1.8.10",
+    "typescript": "^5.7.3"
   }
 }
