@@ -2,17 +2,14 @@ import { Logger } from '@map-colonies/js-logger';
 import { RequestHandler } from 'express';
 import httpStatus from 'http-status-codes';
 import { injectable, inject } from 'tsyringe';
-import { Services } from '../../common/constants';
+import { SERVICES } from '../../common/constants';
 import { ChangeManager } from '../models/changeManager';
-import { ChangeWithMetadata, IdMapping } from '../models/types';
+import { InterpretResult, MergeResult } from '../models/types';
+import { ChangeWithMetadata, OsmXmlChange } from '../models/change';
 
-interface ResponseChangeObject {
-  change: string;
-  created: IdMapping[];
-  deleted: string[];
-}
+type MergeChangesHandler = RequestHandler<undefined, MergeResult, MergeChangesRequestBody>;
 
-type MergeChangesHandler = RequestHandler<undefined, ResponseChangeObject, MergeChangesRequestBody>;
+type InterpretChangeHandler = RequestHandler<undefined, InterpretResult, { osmChange: OsmXmlChange }>;
 
 export interface MergeChangesRequestBody {
   changesetId: number;
@@ -21,10 +18,26 @@ export interface MergeChangesRequestBody {
 
 @injectable()
 export class ChangeController {
-  public constructor(@inject(Services.LOGGER) private readonly logger: Logger, @inject(ChangeManager) private readonly manager: ChangeManager) {}
+  public constructor(
+    @inject(SERVICES.LOGGER) private readonly logger: Logger,
+    @inject(ChangeManager) private readonly manager: ChangeManager
+  ) {}
 
-  public mergeChanges: MergeChangesHandler = (req, res) => {
-    const [change, created, deleted] = this.manager.mergeChanges(req.body.changes, req.body.changesetId);
-    return res.status(httpStatus.OK).json({ change, created, deleted });
+  public mergeChanges: MergeChangesHandler = (req, res, next) => {
+    try {
+      const [change, created, deleted] = this.manager.mergeChanges(req.body.changes, req.body.changesetId);
+      return res.status(httpStatus.OK).json({ change, created, deleted });
+    } catch (error) {
+      return next(error);
+    }
+  };
+
+  public interpretChange: InterpretChangeHandler = (req, res, next) => {
+    try {
+      const result = this.manager.interpretChange(req.body.osmChange);
+      return res.status(httpStatus.OK).json(result);
+    } catch (error) {
+      return next(error);
+    }
   };
 }
