@@ -1,22 +1,30 @@
-import express from 'express';
+import express, { Router } from 'express';
 import bodyParser from 'body-parser';
+import bodyParserXml from 'body-parser-xml';
 import compression from 'compression';
 import { getErrorHandlerMiddleware } from '@map-colonies/error-express-handler';
-import { OpenapiViewerRouter, OpenapiRouterConfig } from '@map-colonies/openapi-express-viewer';
+import { OpenapiViewerRouter } from '@map-colonies/openapi-express-viewer';
 import { middleware as OpenApiMiddleware } from 'express-openapi-validator';
-import { container, inject, injectable } from 'tsyringe';
+import { inject, injectable } from 'tsyringe';
 import { Logger } from '@map-colonies/js-logger';
 import httpLogger from '@map-colonies/express-access-log-middleware';
-import { getTraceContexHeaderMiddleware, defaultMetricsMiddleware } from '@map-colonies/telemetry';
-import { Services } from './common/constants';
-import { IConfig } from './common/interfaces';
-import { changeRouterFactory } from './change/routes/changeRouter';
+import { getTraceContexHeaderMiddleware } from '@map-colonies/telemetry';
+import { Registry } from 'prom-client';
+import { collectMetricsExpressMiddleware } from '@map-colonies/telemetry/prom-metrics';
+import { SERVICES } from './common/constants';
+import { CHANGE_ROUTER_SYMBOL } from './change/routes/changeRouter';
+import { ConfigType } from './common/config';
 
 @injectable()
 export class ServerBuilder {
-  private readonly serverInstance = express();
+  private readonly serverInstance: express.Application;
 
-  public constructor(@inject(Services.CONFIG) private readonly config: IConfig, @inject(Services.LOGGER) private readonly logger: Logger) {
+  public constructor(
+    @inject(SERVICES.CONFIG) private readonly config: ConfigType,
+    @inject(SERVICES.LOGGER) private readonly logger: Logger,
+    @inject(SERVICES.METRICS) private readonly metricsRegistry: Registry,
+    @inject(CHANGE_ROUTER_SYMBOL) private readonly changeRouter: Router
+  ) {
     this.serverInstance = express();
   }
 
@@ -28,29 +36,42 @@ export class ServerBuilder {
     return this.serverInstance;
   }
 
-  private buildRoutes(): void {
-    this.buildDocsRoutes();
-    this.serverInstance.use('/change', changeRouterFactory(container));
+  private buildDocsRoutes(): void {
+    const openapiConfig = this.config.get('openapiConfig');
+    const openapiRouter = new OpenapiViewerRouter({
+      ...openapiConfig,
+      filePathOrSpec: openapiConfig.filePath,
+    });
+    openapiRouter.setup();
+    this.serverInstance.use(openapiConfig.basePath, openapiRouter.getRouter());
   }
 
-  private buildDocsRoutes(): void {
-    const openapiRouter = new OpenapiViewerRouter(this.config.get<OpenapiRouterConfig>('openapiConfig'));
-    openapiRouter.setup();
-    this.serverInstance.use(this.config.get<string>('openapiConfig.basePath'), openapiRouter.getRouter());
+  private buildRoutes(): void {
+    this.serverInstance.use('/change', this.changeRouter);
+    this.buildDocsRoutes();
   }
 
   private registerPreRoutesMiddleware(): void {
-    this.serverInstance.use('/metrics', defaultMetricsMiddleware());
+    this.serverInstance.use(collectMetricsExpressMiddleware({ registry: this.metricsRegistry }));
+    this.serverInstance.use(httpLogger({ logger: this.logger, ignorePaths: ['/metrics'] }));
 
-    if (this.config.get<boolean>('server.response.compression.enabled')) {
-      this.serverInstance.use(compression(this.config.get<compression.CompressionFilter>('server.response.compression.options')));
+    if (this.config.get('server.response.compression.enabled')) {
+      this.serverInstance.use(compression(this.config.get('server.response.compression.options') as unknown as compression.CompressionFilter));
     }
-    this.serverInstance.use(httpLogger({ logger: this.logger }));
-    this.serverInstance.use(express.json(this.config.get<bodyParser.Options>('server.request.payload')));
+
+    const bodyParserOptions = this.config.get('server.request.payload');
+
+    bodyParserXml(bodyParser);
+
+    this.serverInstance.use(bodyParser.json(bodyParserOptions));
+    this.serverInstance.use(
+      bodyParser.xml({ ...bodyParserOptions, type: 'application/xml', xmlParseOptions: { explicitArray: false, mergeAttrs: true, trim: true } })
+    );
+
     this.serverInstance.use(getTraceContexHeaderMiddleware());
 
-    const ignorePathRegex = new RegExp(`^${this.config.get<string>('openapiConfig.basePath')}/.*`, 'i');
-    const apiSpecPath = this.config.get<string>('openapiConfig.filePath');
+    const ignorePathRegex = new RegExp(`^${this.config.get('openapiConfig.basePath')}/.*`, 'i');
+    const apiSpecPath = this.config.get('openapiConfig.filePath');
     this.serverInstance.use(OpenApiMiddleware({ apiSpec: apiSpecPath, validateRequests: true, ignorePaths: ignorePathRegex }));
   }
 
