@@ -1,13 +1,32 @@
-import { Tracing, ignoreIncomingRequestUrl, ignoreOutgoingRequestPath } from '@map-colonies/telemetry';
+import { Tracing, TracingOptions } from '@map-colonies/telemetry';
 import { IGNORED_INCOMING_TRACE_ROUTES, IGNORED_OUTGOING_TRACE_ROUTES } from './constants';
 
-const tracing = new Tracing(undefined, {
-  ['@opentelemetry/instrumentation-http']: {
-    ignoreIncomingRequestHook: ignoreIncomingRequestUrl(IGNORED_INCOMING_TRACE_ROUTES),
-    ignoreOutgoingRequestHook: ignoreOutgoingRequestPath(IGNORED_OUTGOING_TRACE_ROUTES),
-  },
-});
+let tracing: Tracing | undefined;
 
-tracing.start();
+export function tracingFactory(options: TracingOptions): Tracing {
+  tracing = new Tracing({
+    ...options,
+    autoInstrumentationsConfigMap: {
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      '@opentelemetry/instrumentation-http': {
+        ignoreIncomingRequestHook: (request): boolean =>
+          IGNORED_INCOMING_TRACE_ROUTES.some((route) => request.url !== undefined && route.test(request.url)),
+        ignoreOutgoingRequestHook: (request): boolean =>
+          IGNORED_OUTGOING_TRACE_ROUTES.some((route) => typeof request.path === 'string' && route.test(request.path)),
+      },
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      '@opentelemetry/instrumentation-fs': {
+        requireParentSpan: true,
+      },
+    },
+  });
 
-export { tracing };
+  return tracing;
+}
+
+export function getTracing(): Tracing {
+  if (!tracing) {
+    throw new Error('tracing not initialized');
+  }
+  return tracing;
+}
