@@ -1,31 +1,85 @@
-import { container } from 'tsyringe';
-import config from 'config';
-import { Metrics, getOtelMixin } from '@map-colonies/telemetry';
-import { metrics } from '@opentelemetry/api-metrics';
-import jsLogger, { LoggerOptions } from '@map-colonies/js-logger';
+import { getOtelMixin } from '@map-colonies/telemetry';
 import { trace } from '@opentelemetry/api';
-import { tracing } from './common/tracing';
-import { Services } from './common/constants';
+import { Registry } from 'prom-client';
+import { DependencyContainer } from 'tsyringe/dist/typings/types';
+import jsLogger, { Logger } from '@map-colonies/js-logger';
+import { InjectionObject, registerDependencies } from '@common/dependencyRegistration';
+import { instancePerContainerCachingFactory } from 'tsyringe';
+import { CleanupRegistry } from '@map-colonies/cleanup-registry';
+import { HEALTHCHECK, ON_SIGNAL, SERVICES, SERVICE_NAME, stubHealthCheck } from '@common/constants';
+import { getTracing } from '@common/tracing';
+import { ConfigType, getConfig } from './common/config';
+import { CHANGE_ROUTER_SYMBOL, changeRouterFactory } from './change/routes/changeRouter';
 
-function registerExternalValues(): void {
-  const loggerConfig = config.get<LoggerOptions>('telemetry.logger');
-  const logger = jsLogger({ ...loggerConfig, prettyPrint: loggerConfig.prettyPrint, mixin: getOtelMixin() });
-
-  container.register(Services.CONFIG, { useValue: config });
-  container.register(Services.LOGGER, { useValue: logger });
-
-  const tracer = trace.getTracer('change-merger');
-  container.register(Services.TRACER, { useValue: tracer });
-
-  const otelMetrics = new Metrics();
-  otelMetrics.start();
-  container.register(Services.METER, { useValue: metrics.getMeter('change-merger') });
-
-  container.register('onSignal', {
-    useValue: async (): Promise<void> => {
-      await Promise.all([tracing.stop(), otelMetrics.stop()]);
-    },
-  });
+export interface RegisterOptions {
+  override?: InjectionObject<unknown>[];
+  useChild?: boolean;
 }
 
-export { registerExternalValues };
+export const registerExternalValues = async (options?: RegisterOptions): Promise<DependencyContainer> => {
+  const cleanupRegistry = new CleanupRegistry();
+
+  try {
+    const configInstance = getConfig();
+    const loggerConfig = configInstance.get('telemetry.logger');
+    const logger = jsLogger({ ...loggerConfig, mixin: getOtelMixin() });
+
+    const dependencies: InjectionObject<unknown>[] = [
+      { token: SERVICES.CONFIG, provider: { useValue: configInstance } },
+      { token: SERVICES.LOGGER, provider: { useValue: logger } },
+      {
+        token: SERVICES.CLEANUP_REGISTRY,
+        provider: { useValue: cleanupRegistry },
+        postInjectionHook(container): void {
+          const logger = container.resolve<Logger>(SERVICES.LOGGER);
+          const cleanupRegistryLogger = logger.child({ subComponent: 'cleanupRegistry' });
+
+          cleanupRegistry.on('itemFailed', (id, error, msg) => cleanupRegistryLogger.error({ msg, itemId: id, err: error }));
+          cleanupRegistry.on('itemCompleted', (id) => cleanupRegistryLogger.info({ itemId: id, msg: 'cleanup finished for item' }));
+          cleanupRegistry.on('finished', (status) => cleanupRegistryLogger.info({ msg: `cleanup registry finished cleanup`, status }));
+        },
+      },
+      {
+        token: SERVICES.TRACER,
+        provider: {
+          useFactory: instancePerContainerCachingFactory((container) => {
+            const cleanupRegistry = container.resolve<CleanupRegistry>(SERVICES.CLEANUP_REGISTRY);
+            cleanupRegistry.register({ id: SERVICES.TRACER, func: getTracing().stop.bind(getTracing()) });
+            const tracer = trace.getTracer(SERVICE_NAME);
+            return tracer;
+          }),
+        },
+      },
+      {
+        token: SERVICES.METRICS,
+        provider: {
+          useFactory: instancePerContainerCachingFactory((container) => {
+            const metricsRegistry = new Registry();
+            const config = container.resolve<ConfigType>(SERVICES.CONFIG);
+            config.initializeMetrics(metricsRegistry);
+            return metricsRegistry;
+          }),
+        },
+      },
+      { token: CHANGE_ROUTER_SYMBOL, provider: { useFactory: changeRouterFactory } },
+      {
+        token: HEALTHCHECK,
+        provider: {
+          useValue: stubHealthCheck,
+        },
+      },
+      {
+        token: ON_SIGNAL,
+        provider: {
+          useValue: cleanupRegistry.trigger.bind(cleanupRegistry),
+        },
+      },
+    ];
+
+    const container = await registerDependencies(dependencies, options?.override, options?.useChild);
+    return container;
+  } catch (error) {
+    await cleanupRegistry.trigger();
+    throw error;
+  }
+};
