@@ -1,25 +1,30 @@
+const { pathsToModuleNameMapper } = require('ts-jest');
+const { compilerOptions } = require('../../../tsconfig.json');
+
+/** @type {import('jest').Config} */
 module.exports = {
   transform: {
-    '^.+\\.ts$': 'ts-jest',
+    '^.+\\.ts$': ['@swc/jest'],
   },
+  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, { prefix: '<rootDir>/' }),
   testMatch: ['<rootDir>/tests/unit/**/*.spec.ts'],
+  coverageReporters: ['text', 'html'],
+  collectCoverage: true,
+  collectCoverageFrom: ['<rootDir>/src/change/models/**/*.ts', '!**/node_modules/**', '!**/vendor/**'],
+  coverageDirectory: '<rootDir>/coverage',
   reporters: [
     'default',
     ['jest-html-reporters', { multipleReportsUnitePath: './reports', pageTitle: 'unit', publicPath: './reports', filename: 'unit.html' }],
   ],
   rootDir: '../../../.',
-  setupFiles: ['<rootDir>/tests/configurations/jest.setup.js'],
-  preset: 'ts-jest',
+  setupFiles: ['<rootDir>/tests/configurations/jest.setup.ts'],
   testEnvironment: 'node',
-  collectCoverage: true,
-  collectCoverageFrom: ['<rootDir>/src/change/models/**/*.ts', '!**/node_modules/**', '!**/vendor/**'],
-  coverageReporters: ['text', 'html', 'json'],
   coverageThreshold: {
     global: {
       branches: 80,
       functions: 80,
       lines: 80,
-      statments: -10,
+      statements: -10,
     },
   },
 };
